requirements.txt

# Web framework
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0

# OCR
python-doctr==0.7.0
doctr[torch]==0.7.0

# Image processing
Pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.24.3

# Machine learning
scikit-learn==1.3.0
pytorch-lightning==2.0.6
torch==2.0.1
torchvision==0.15.2
transformers==4.32.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
